make webpack config
