+++ Gruntfile.js
@@ -22,6 +22,7 @@ module.exports = function(grunt) {
         sass: 'grunt-sass',
         connect: 'grunt-contrib-connect',
         assemble: 'grunt/assemble/',
+        webpack: 'grunt/webpack/',
         mochaTest: 'grunt-mocha-test',
         open: 'grunt-open'
       }
@@ -40,13 +41,11 @@ module.exports = function(grunt) {
   grunt.registerTask('staging-deploy', [
     'gitinfo',
     'config:staging',
-    'jshint:clientDev',
-    'jshint:server',
     'clean:preBuild',
     'assemble',
-    'handlebars',
     'modernizr',
     'concat',
+    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -58,13 +57,11 @@ module.exports = function(grunt) {
   grunt.registerTask('smartling-staging-deploy', [
     'gitinfo',
     'config:smartlingStaging',
-    'jshint:clientDev',
-    'jshint:server',
     'clean:preBuild',
     'assemble:smartling-staging-deploy',
-    'handlebars',
     'modernizr',
     'concat',
+    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -75,14 +72,12 @@ module.exports = function(grunt) {
 
   var serverTasks = [
     'config:dev',
-    'jshint:clientDev',
-    'jshint:server',
     'jshint:test',
     'clean:preBuild',
     'assemble',
-    'handlebars',
     'modernizr',
     'concat',
+    'webpack',
     'sass:dev',
     'replace',
     'autoprefixer',
@@ -111,13 +106,11 @@ module.exports = function(grunt) {
 
   grunt.registerTask('build', [
     'config:production',
-    'jshint:clientProd',
-    'jshint:server',
     'clean:preBuild',
     'assemble',
-    'handlebars',
     'modernizr',
     'concat',
+    'webpack',
     'sass:prod',
     'autoprefixer',
     'copy',
@@ -151,14 +144,12 @@ module.exports = function(grunt) {
 
   grunt.registerTask('test', [
     'config:dev',
-    'jshint:clientProd',
-    'jshint:server',
     'jshint:test',
     'clean:preBuild',
     'assemble',
-    'handlebars',
     'modernizr',
     'concat',
+    'webpack',
     'sass:dev',
     'replace',
     'autoprefixer',
