+++ grunt/webpack/webpack.config.js
@@ -1,0 +1,3 @@
+module.exports = require("./make-webpack-config")({
+
+});
