+++ package.json
@@ -23,6 +23,9 @@
   "dependencies": {
     "assemble": "0.6.0-beta.5",
     "assemble-push": "^0.1.1",
+    "babel-core": "^5.0.12",
+    "babel-loader": "^5.0.0",
+    "babel-runtime": "^5.0.12",
     "body-parser": "1.8.3",
     "bower": "1.3.8",
     "chalk": "1.0.0",
@@ -36,7 +39,6 @@
     "grunt-contrib-clean": "0.5.0",
     "grunt-contrib-concat": "0.4.0",
     "grunt-contrib-copy": "0.5.0",
-    "grunt-contrib-handlebars": "0.8.0",
     "grunt-contrib-imagemin": "0.7.1",
     "grunt-contrib-jshint": "0.10.0",
     "grunt-contrib-uglify": "0.5.0",
@@ -48,18 +50,26 @@
     "grunt-text-replace": "0.3.12",
     "grunt-userevvd": "0.1.4",
     "gulp-extname": "0.2.0",
-    "handlebars": "3.0.0",
+    "gulp-util": "^3.0.4",
+    "handlebars-loader": "0.1.7",
+    "imports-loader": "0.6.3",
+    "inject-filename-loader": "git+ssh://git@github.com:dtothefp/inject-filename-loader.git",
+    "jshint-loader": "0.8.1",
     "l10n-tools": "git+ssh://git@github.com/optimizely/l10n-tools.git",
     "layout-stack": "^0.1.1",
     "load-grunt-config": "0.16.0",
     "lodash": "^3.2.0",
     "mkdirp": "0.5.0",
+    "moment": "^2.10.2",
     "optimizely-marked": "0.0.4",
     "plasma": "^0.6.1",
     "q": "1.1.2",
+    "script-loader": "0.6.1",
     "through2": "0.6.3",
     "time-grunt": "0.3.2",
-    "uglify-js": "2.4.15"
+    "uglify-js": "2.4.15",
+    "webpack": "^1.8.4",
+    "webpack-dev-server": "^1.8.0"
   },
   "devDependencies": {
     "chai": "2.2.0",
