+++ website-guts/assets/js/components/utility-nav.js
@@ -1,4 +1,5 @@
 var lastDropdown;
+var loggedInUtilityNav = require('loggedInUtilityNav');
 
 function bindDropdownClick($dropdownMenus) {
 
@@ -68,7 +69,7 @@ window.optly.mrkt.showUtilityNav = function (acctData, expData) {
 
     $('body').addClass('signed-in').removeClass('signed-out');
 
-    $('#signed-in-utility').html( window.optly.mrkt.templates.loggedInUtilityNav(handlebarsData) );
+    $('#signed-in-utility').html( loggedInUtilityNav(handlebarsData) );
     var $dropdownMenus = $('[data-show-dropdown]');
 
     bindDropdownClick($dropdownMenus);
