+++ website-guts/assets/js/pages/events.js
@@ -1,3 +1,6 @@
+var moment = require('moment');
+var eventDisplay = require('eventDisplay');
+
 window.optly.mrkt.events = {};
 
 var gapi = window.gapi;
@@ -28,8 +31,8 @@ window.optly.mrkt.events.showEvents = function(resp, div){
       var entry, eventData, venue, startDate, endDate, zeroRegEx, description;
 
       entry = items[i];
-      startDate = window.moment( entry.start.date || entry.start.dateTime );
-      endDate = window.moment( entry.end.date || entry.end.dateTime );
+      startDate = moment( entry.start.date || entry.start.dateTime );
+      endDate = moment( entry.end.date || entry.end.dateTime );
       zeroRegEx = /\-0/g;
 
       if(typeof entry.location === 'string'){
@@ -69,7 +72,7 @@ window.optly.mrkt.events.showEvents = function(resp, div){
       pastEvents.splice(0, pastEvents.length - 30);
     }
 
-    $(div).html(window.optly.mrkt.templates.eventDisplay({events: futureEvents}));
+    $(div).html(eventDisplay({events: futureEvents}));
 
     $('#get-past-events').on('click', function(e){
       e.preventDefault();
@@ -82,7 +85,7 @@ window.optly.mrkt.events.showEvents = function(resp, div){
           $(elm).removeClass('hide-past').addClass('hide-upcoming');
         }
       });
-      $(div).html(window.optly.mrkt.templates.eventDisplay({events: pastEvents}));
+      $(div).html(eventDisplay({events: pastEvents}));
     });
 
     $('#get-future-events').on('click', function(e){
@@ -96,7 +99,7 @@ window.optly.mrkt.events.showEvents = function(resp, div){
           $(elm).removeClass('hide-upcoming').addClass('hide-past');
         }
       });
-      $(div).html(window.optly.mrkt.templates.eventDisplay({events: futureEvents}));
+      $(div).html(eventDisplay({events: futureEvents}));
     });
 
   } else {
