+++ website-guts/assets/js/pages/jobs.js
@@ -1,4 +1,5 @@
 window.optly.mrkt.jobsPage = {};
+var jobList = require('jobList');
 
 window.optly.mrkt.jobsPage.testimonials = function() {
     var lastIndex = 0;
@@ -37,7 +38,7 @@ function getGreenhouseData(data) {
           }
         }
 
-        $('#job-list-cont').append( window.optly.mrkt.templates.jobList(data) );
+        $('#job-list-cont').append( jobList(data) );
       }
 }
